chore: webpack\rsbuild\vite 都使用 core 统一简洁输出和增量写入
